requirements.txt
flask==3.0.3
flask-cors==4.0.1
reportlab==4.2.2
stripe==10.6.0
google-api-python-client==2.131.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
requests==2.32.3
gunicorn==22.0.0